Fix demo tests
