+++ demos/dqx_demo_tool.py
@@ -53,9 +53,11 @@ import os
 
 user_name = spark.sql("select current_user() as user").collect()[0]["user"]
 default_dqx_wheel_files = glob.glob(f"/Workspace/Users/{user_name}/.dqx/wheels/databricks_labs_dqx-*.whl")
+default_dqx_wheel_files_path = f"/Workspace/Users/{user_name}/.dqx/wheels/databricks_labs_dqx-*.whl"
 
-dbutils.widgets.text("dqx_wheel_files", default_dqx_wheel_files, "DQX Wheel Files Folder")
-dqx_wheel_files = dbutils.widgets.get("dqx_wheel_files")
+dbutils.widgets.text("dqx_wheel_files_path", default_dqx_wheel_files_path, "DQX Wheel Files Folder")
+dqx_wheel_files_path = dbutils.widgets.get("dqx_wheel_files_path")
+dqx_wheel_files = glob.glob(dqx_wheel_files_path)
 
 dqx_latest_wheel = max(dqx_wheel_files, key=os.path.getctime)
 %pip install {dqx_latest_wheel}
